LeastSquaresProblem moved to solve
