+++ examples/just_a_quadratic
@@ -2,7 +2,7 @@
 
 import logging
 from simsopt.core.functions import Identity
-from simsopt.core.least_squares_problem import LeastSquaresProblem
+from simsopt.solve.least_squares_problem import LeastSquaresProblem
 from simsopt.solve.serial_solve import least_squares_serial_solve
 
 """
