+++ examples/qsc
@@ -4,7 +4,7 @@ import logging
 import numpy as np
 from qsc import Qsc
 from simsopt.core.optimizable import optimizable
-from simsopt.core.least_squares_problem import LeastSquaresProblem
+from simsopt.solve.least_squares_problem import LeastSquaresProblem
 from simsopt.solve.serial_solve import least_squares_serial_solve
 
 """
