+++ examples/stellopt_scenarios_1DOF_circularCrossSection_varyAxis_targetIota
@@ -4,7 +4,7 @@ from mpi4py import MPI
 import numpy as np
 
 from simsopt.mhd import Vmec
-from simsopt.core.least_squares_problem import LeastSquaresProblem
+from simsopt.solve.least_squares_problem import LeastSquaresProblem
 from simsopt.util.mpi import MpiPartition, log
 from simsopt.solve.mpi_solve import least_squares_mpi_solve
