+++ examples/stellopt_scenarios_1DOF_circularCrossSection_varyAxis_targetIota_spec
@@ -4,7 +4,7 @@ import logging
 import numpy as np
 
 from simsopt.mhd import Spec
-from simsopt.core.least_squares_problem import LeastSquaresProblem
+from simsopt.solve.least_squares_problem import LeastSquaresProblem
 from simsopt.solve.serial_solve import least_squares_serial_solve
 
 """
