+++ examples/stellopt_scenarios_1DOF_circularCrossSection_varyR0_targetVolume
@@ -3,7 +3,7 @@
 import numpy as np
 from mpi4py import MPI
 
-from simsopt.core.least_squares_problem import LeastSquaresProblem
+from simsopt.solve.least_squares_problem import LeastSquaresProblem
 from simsopt.mhd import Vmec
 from simsopt.util.mpi import MpiPartition, log
 from simsopt.solve.mpi_solve import least_squares_mpi_solve
