+++ examples/stellopt_scenarios_2DOF_vmecAndSpec
@@ -4,7 +4,7 @@ import logging
 import numpy as np
 
 from simsopt.mhd import Vmec, Spec
-from simsopt.core.least_squares_problem import LeastSquaresProblem
+from simsopt.solve.least_squares_problem import LeastSquaresProblem
 from simsopt.solve.serial_solve import least_squares_serial_solve
 
 """
