+++ examples/stellopt_scenarios_2DOF_vmecOnly_targetIotaAndVolume
@@ -3,7 +3,7 @@
 import numpy as np
 
 from simsopt.mhd import Vmec
-from simsopt.core.least_squares_problem import LeastSquaresProblem
+from simsopt.solve.least_squares_problem import LeastSquaresProblem
 from simsopt.util.mpi import MpiPartition, log
 from simsopt.solve.mpi_solve import least_squares_mpi_solve
