+++ examples/surface_volume_and_area
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 
 from simsopt.geo.surfacerzfourier import SurfaceRZFourier
-from simsopt.core.least_squares_problem import LeastSquaresProblem
+from simsopt.solve.least_squares_problem import LeastSquaresProblem
 from simsopt.solve.serial_solve import least_squares_serial_solve
 """
 Optimize the minor radius and elongation of an axisymmetric torus to
