+++ tests/core/test_integrated.py
@@ -6,8 +6,8 @@ import numpy as np
 from simsopt.geo.surfacerzfourier import SurfaceRZFourier
 from simsopt.geo.surfacegarabedian import SurfaceGarabedian
 from simsopt.core.optimizable import optimizable
-from simsopt.core.least_squares_problem import LeastSquaresProblem
 from simsopt.util.mpi import MpiPartition
+from simsopt.solve.least_squares_problem import LeastSquaresProblem
 from simsopt.solve.serial_solve import least_squares_serial_solve
 from simsopt.solve.mpi_solve import least_squares_mpi_solve
