+++ tests/geo/test_curve_optimizable.py
@@ -7,7 +7,7 @@ from simsopt.geo.curve import RotatedCurve
 from simsopt.geo import parameters
 from simsopt.geo.curveobjectives import CurveLength
 from simsopt.core.optimizable import optimizable
-from simsopt.core.least_squares_problem import LeastSquaresProblem
+from simsopt.solve.least_squares_problem import LeastSquaresProblem
 from simsopt.solve.serial_solve import least_squares_serial_solve
 
 parameters['jit'] = False
