+++ tests/mhd/test_integrated_vmec_mpi.py
@@ -4,8 +4,8 @@ import logging
 import numpy as np
 from mpi4py import MPI
 
-from simsopt.core.least_squares_problem import LeastSquaresProblem
 from simsopt.mhd.vmec import Vmec, vmec_found
+from simsopt.solve.least_squares_problem import LeastSquaresProblem
 from simsopt.solve.mpi_solve import least_squares_mpi_solve
 from simsopt.util.mpi import MpiPartition
