+++ tests/mhd/test_spec.py
@@ -4,7 +4,7 @@ import os
 import logging
 import shutil
 from simsopt.mhd.spec import Spec, nested_lists_to_array
-from simsopt.core.least_squares_problem import LeastSquaresProblem
+from simsopt.solve.least_squares_problem import LeastSquaresProblem
 from simsopt.solve.serial_solve import least_squares_serial_solve
 from . import TEST_DIR
