+++ tests/solve/test_least_squares_solve.py
@@ -3,7 +3,7 @@ import logging
 
 from simsopt.core.functions import Identity, Rosenbrock
 from simsopt.core.optimizable import Target
-from simsopt.core.least_squares_problem import LeastSquaresProblem, LeastSquaresTerm
+from simsopt.solve.least_squares_problem import LeastSquaresProblem, LeastSquaresTerm
 from simsopt.solve.serial_solve import least_squares_serial_solve
 from simsopt.util.mpi import MpiPartition
 from simsopt.solve.mpi_solve import least_squares_mpi_solve
