+++ tests/solve/test_mpi.py
@@ -3,8 +3,8 @@ import unittest
 import numpy as np
 from mpi4py import MPI
 from simsopt.core.dofs import Dofs
-from simsopt.core.least_squares_problem import LeastSquaresProblem
 from simsopt.util.mpi import MpiPartition
+from simsopt.solve.least_squares_problem import LeastSquaresProblem
 from simsopt.solve.mpi_solve import fd_jac_mpi, least_squares_mpi_solve
 
 #logging.basicConfig(level=logging.DEBUG)
